Revert "leveraging CI to run a CTK 12.1 test"
This reverts commit a94e41592a0653da8365ae2c6b72ba9697270050.
