+++ .github/workflows/pr.yaml
@@ -25,36 +25,25 @@ jobs:
     uses: rapidsai/shared-workflows/.github/workflows/checks.yaml@branch-24.04
     with:
       enable_check_generated_files: false
-  # legacy_tests:
-  #   name: Runs on legacy hardware and older CTK versions (V100, 12.3)
-  #   needs: checks
-  #   secrets: inherit
-  #   uses: ./.github/workflows/setup_build_test.yaml
-  #   with:
-  #     ubuntu_version: "20.04"
-  #     cuda_version: "12.3"
-  #     python_version: "3.10"
-  #     GPU: "v100"
-  #     CPU: "amd64"
-  # frontier_tests:
-  #   name: Runs on newer hardware and latest CTK versions (A100, 12.4)
-  #   needs: checks
-  #   secrets: inherit
-  #   uses: ./.github/workflows/setup_build_test.yaml
-  #   with:
-  #     ubuntu_version: "22.04"
-  #     cuda_version: "12.4"
-  #     python_version: "3.10"
-  #     GPU: "a100"
-  #     CPU: "arm64"
-  pytorch_tests:
-    name: Temp job on pytorch
+  legacy_tests:
+    name: Runs on legacy hardware and older CTK versions (V100, 12.3)
+    needs: checks
+    secrets: inherit
+    uses: ./.github/workflows/setup_build_test.yaml
+    with:
+      ubuntu_version: "20.04"
+      cuda_version: "12.3"
+      python_version: "3.10"
+      GPU: "v100"
+      CPU: "amd64"
+  frontier_tests:
+    name: Runs on newer hardware and latest CTK versions (A100, 12.4)
     needs: checks
     secrets: inherit
     uses: ./.github/workflows/setup_build_test.yaml
     with:
       ubuntu_version: "22.04"
-      cuda_version: "12.1"
+      cuda_version: "12.4"
       python_version: "3.10"
       GPU: "a100"
       CPU: "arm64"
