+++ numba_extensions/tests/test_curand.py
@@ -34,9 +34,6 @@ from curand_host import (
     curandDirectionVectorSet,
 )
 
-
-pytest.skip(reason="skip until test is fixed", allow_module_level=True)
-
 # Various parameters
 
 threads = 64
