+++ numba_extensions/tests/test_fp162.py
@@ -26,10 +26,6 @@ import numpy as np
 from numba import float32
 import numba.cuda as cuda
 
-import pytest
-
-pytest.skip(reason="skip until test is fixed", allow_module_level=True)
-
 
 def test_ctor():
     @cuda.jit(link=get_shims())
